+++ end_to_end_tests/__init__.py
@@ -2,3 +2,4 @@
 import pytest
 
 pytest.register_assert_rewrite("end_to_end_tests.end_to_end_test_helpers")
+pytest.register_assert_rewrite("end_to_end_tests.functional_tests.helpers")
