+++ end_to_end_tests/functional_tests/generated_code_execution/test_arrays.py
@@ -1,8 +1,6 @@
-import datetime
 from typing import Any, ForwardRef, List, Union
-import uuid
-import pytest
-from end_to_end_tests.end_to_end_test_helpers import (
+
+from end_to_end_tests.functional_tests.helpers import (
     assert_model_decode_encode,
     assert_model_property_type_hint,
     with_generated_client_fixture,
