+++ end_to_end_tests/functional_tests/generated_code_execution/test_defaults.py
@@ -1,6 +1,7 @@
 import datetime
 import uuid
-from end_to_end_tests.end_to_end_test_helpers import (
+
+from end_to_end_tests.functional_tests.helpers import (
     with_generated_client_fixture,
     with_generated_code_imports,
 )
