+++ end_to_end_tests/functional_tests/generator_failure_cases/test_invalid_arrays.py
@@ -1,14 +1,9 @@
 import pytest
-from end_to_end_tests.end_to_end_test_helpers import (
-    assert_bad_schema_warning,
-    inline_spec_should_cause_warnings,
-)
+
+from end_to_end_tests.functional_tests.helpers import assert_bad_schema, with_generated_client_fixture
 
 
-class TestArrayInvalidSchemas:
-    @pytest.fixture(scope="class")
-    def warnings(self):
-        return inline_spec_should_cause_warnings(
+@with_generated_client_fixture(
 """
 components:
   schemas:
@@ -18,11 +13,11 @@ components:
       type: array
       items:
         $ref: "#/components/schemas/DoesntExist"
-"""
-        )
+"""    
+)
+class TestArrayInvalidSchemas:
+    def test_no_items(self, generated_client):
+        assert_bad_schema(generated_client, "ArrayWithNoItems", "must have items or prefixItems defined")
 
-    def test_no_items(self, warnings):
-        assert_bad_schema_warning(warnings, "ArrayWithNoItems", "must have items or prefixItems defined")
-
-    def test_invalid_items_ref(self, warnings):
-        assert_bad_schema_warning(warnings, "ArrayWithInvalidItemsRef", "invalid data in items of array")
+    def test_invalid_items_ref(self, generated_client):
+        assert_bad_schema(generated_client, "ArrayWithInvalidItemsRef", "invalid data in items of array")
