+++ end_to_end_tests/functional_tests/generator_failure_cases/test_invalid_unions.py
@@ -1,14 +1,7 @@
-import pytest
-from end_to_end_tests.end_to_end_test_helpers import (
-    assert_bad_schema_warning,
-    inline_spec_should_cause_warnings,
-)
+from end_to_end_tests.functional_tests.helpers import assert_bad_schema, with_generated_client_fixture
 
 
-class TestUnionInvalidSchemas:
-    @pytest.fixture(scope="class")
-    def warnings(self):
-        return inline_spec_should_cause_warnings(
+@with_generated_client_fixture(
 """
 components:
   schemas:
@@ -22,14 +15,14 @@ components:
       anyOf:
         - type: string
         - type: array  # invalid because no items
-"""
-        )
+"""    
+)
+class TestUnionInvalidSchemas:
+    def test_invalid_reference(self, generated_client):
+        assert_bad_schema(generated_client, "UnionWithInvalidReference", "Could not find reference")
 
-    def test_invalid_reference(self, warnings):
-        assert_bad_schema_warning(warnings, "UnionWithInvalidReference", "Could not find reference")
+    def test_invalid_default(self, generated_client):
+        assert_bad_schema(generated_client, "UnionWithInvalidDefault", "Invalid int value: aaa")
 
-    def test_invalid_default(self, warnings):
-        assert_bad_schema_warning(warnings, "UnionWithInvalidDefault", "Invalid int value: aaa")
-
-    def test_invalid_property(self, warnings):
-        assert_bad_schema_warning(warnings, "UnionWithMalformedVariant", "Invalid property in union")
+    def test_invalid_property(self, generated_client):
+        assert_bad_schema(generated_client, "UnionWithMalformedVariant", "Invalid property in union")
